ext4 crypto: fix return value for ext4_es_scan()
Between 3.10 and 3.18, the abstraction to scan for objects in the slab
cache which can be freed when the system is under memory pressure
changed.  When I backported the ext4 code from 3.18 to the 3.10
kernel, I didn't get the return value required by the calling
conventions for the scan function correct, which could potentially
cause the memory reclaimer to loop indefinitely.

Change-Id: I1712fedf96fd91c911fb4d019d7ef76f6c4c1808
Signed-off-by: "Theodore Ts'o" <tytso@google.com>
Signed-off-by: AudioGod <audiogod@sonic-developers.com>